Get test for Sum of Squares working
Adding up the sum of the numbers up to where the loop equals the number the user has given. Once the loop has finished we need to square the number, cast to an int and return it to the user.

The other million tests are failing, but that is tomorrows job.